Move ExportExamplesButton component to elements subfolder
