+++ app/packs/src/components/elements/wellplates/WellplateDetails.js
@@ -22,7 +22,7 @@ import Utils from 'src/utilities/Functions';
 import UIStore from 'src/stores/alt/stores/UIStore';
 import UIActions from 'src/stores/alt/actions/UIActions';
 import ConfirmClose from 'src/components/common/ConfirmClose';
-import ExportSamplesBtn from 'src/components/ExportSamplesBtn';
+import ExportSamplesBtn from 'src/components/elements/ExportSamplesBtn';
 import ElementDetailSortTab from 'src/components/elements/ElementDetailSortTab';
 import { addSegmentTabs } from 'src/components/generic/SegmentDetails';
 import PrivateNoteElement from 'src/components/PrivateNoteElement'
